fix(app): move cloud sign-in CTA to header
